[refactor] #25 Endpoint 각도메인별로 분리위해 protocol 추상화
