+++ Logit/Logit/Network/Core/EndPoint/HTTPMethod.swift
@@ -1,0 +1,16 @@
+//
+//  HTTPMethod.swift
+//  Logit
+//
+//  Created by 임재현 on 1/30/26.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case get = "GET"
+    case post = "POST"
+    case put = "PUT"
+    case delete = "DELETE"
+    case patch = "PATCH"
+}
