+++ Logit/Logit/Network/Core/NetworkClient.swift
@@ -9,13 +9,13 @@ import Foundation
 
 protocol NetworkClient {
     func request<T:Decodable>(
-        endpoint: APIEndpoint,
+        endpoint: AuthEndpoint,
         body: Encodable?
     ) async throws -> T
     
     // 응답 body 없을때
     func request(
-        endpoint: APIEndpoint,
+        endpoint: AuthEndpoint,
         body: Encodable?
     ) async throws
 }
